Remove prange utils to other file
